requirements.txt
beautifulsoup4
mastodon.py
